server/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
pymysql
pymysql
python-dotenv
passlib[bcrypt]
python-jose
paho-mqtt>=2.0.0
